sq/backend/requirements.txt
click==8.1.3
dnspython==2.2.1
flask-cors==3.0.10
flask-pymongo==2.3.0
flask==2.2.2
importlib-metadata==5.1.0; python_version < '3.10'
itsdangerous==2.1.2
jinja2==3.1.2
markupsafe==2.1.1
pymongo==4.3.3
six==1.16.0
werkzeug==2.2.2
zipp==3.11.0
PyJWT==2.8.0
PyCryptodome==3.19.0
gunicorn
